requirements.txt
requests
python-dotenv
Pillow
openai
tweepy 
schedule